Added registration functionality and new messages.
Registry can take multiple connections and register the MessagingNodes.
The messaging nodes are sending their request messages and the registry
is reading and doing the reponses asked of it. However, MessagingNodes
are not receiving the messages sent by the registry quite yet. Still
need to do that. 

TODO: Create my own version of the TCPReceiver to handle messages sent to messaging nodes itself.