Registry and Messaging Node exchanging messages
Right now I have the registry and the messaging nodes exchanging
messages. At the current point, the messaging nodes are not doing
anything with the message that was received from the registry.

What to do next:
- Get a few nodes connected to the registry and then send one of the
messaging nodes an IP address and port number and have one of the nodes
connect to another node.
- Once nodes are connected to each other, have them send messages.
- Try getting around 10 different node connections. If it works to
somewhat of some degree, then we have completed milestone 2.

* add some more of the registry commands.
* add functionality to communicate with other nodes in the system from
within the MessagingNode.