Added new Thread for listening to user inputs.
These threads are working correctly right now. They take input from the
user and have an instance of the class calling them. The only thing that
needs to be done next is to send the messages back to the node calling
them. For right now, it is not important.

Next its time to update how to do the overlay. Need to build a graph.
Then run Dijkstra's algorithm. Then set up the overlay. Then we can go
about sending messages to the nodes and having them set up connections.