Added Graph representation of MessagingNodes in the overlay.
The Registry will keep track of current port numbers and IP addresses,
and is in charge of adding nodes to the graph overlay. 

Still have to build dijkstra's algorithm to find the shortest path.
Next, I will also have to add in the rest of the commands and get all
messaging nodes to connect to where they are suppose to connect to.