Added some fixes and identified problems.
Unfortunately there is a major logic problem going on in my code. I
cannot get it figured out for the life of me. I need to look through
topological sort again because for whatever reason it worked at first
and it doesn't now. I need to figure out why it is not working anymore.
However, I do have edges and assigned weights being attached to them. 

Once I get the sort finished I begin testing Dijkstra's algorithm and
once that is working the rest of the project will be pretty much easy.