Dijkstra's alg is working and we have edge weights being assigned.
There is also the massive MessagingNodesList class that is built as
well. This is a relatively simple class but much still has to be done on
it. There are some indexing problems that I need to get addressed before
I can progress through, but once this is working I can assign the
message list to the MessagingNodes and send this bad boy off to be
processed by each of the nodes.

Once that is done, I must have all the nodes complete their connections.

Once all connections are complete, I can then clean my code up before
proceeding to the next step to make sure I know what I'm doing.