Added class path and .project for campus work. Nodes get node list now.
I have made it so that all messaging nodes are now getting the proper
amount of data needed to make connections with the other nodes. It is
now time to make sure that we make connections with the node information
that was sent to the other nodes. 

After that, we can begin to use Dijkstra's alg and begin to set up the
payload to send to the other nodes. 

The nodes should know which node that this payload should reach so we
should make sure to get started on this bad boy and make sure that we
get started on this bad boy and start killing it. 

Long sleepless nights await me.