Clean up, Added new template class, created LinkWeightMessage
I created a new template class to help with maintaining the edges while
writing the LinkWeightMessage. This makes things a lot easier and
cleaner. I also cleaned up a few things in the Dijkstra class. I also
added the command "send-overlay-link-weights" to the registry.

Still have to have the messaging nodes process this message now. I also
have to add some other messages and try to send the payload to the
correct nodes and make sure that I am keeping track of the data that is
being sent. This is very important.