Finished adding commands and beginning to have nodes start their rounds.
This project is almost completed. I just have to make sure that I am
getting the shortest path from Dijkstra's algorithm correctly. This is
crucial. This means I have to ask the register to send me that path so
that the nodes should know how to send the data. The nodes will have to
send their data to one of the nodes that they are connected, it's just a
matter of how many nodes the message must get sent through in order for
it to reach it's destination. The destination is the node that we send
to the Registry in order to compute the shortest path for it.

Read my notes on how I went about sending this data.

This is done once per round for each messaging node.