Getting messaging nodes the shortest path list.
The last hurdle I have to get through is to get the messaging nodes
sending data the way that they should be. In order to do that, I need to
setup a shortestPathList, this is very important.

If I continue to run into the problem where the nodesList is coming up
null, I may have to apply the wait and notify trick to get this thing to
read in the proper way that it should. This is not going to be easy, but
once I get this working there is literally nothing else I cannot do. I
just gotta make sure the messaging nodes are able to send messages
between each other and update the correct fields.